pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "web-auth-mcp"
version = "0.1.0"
description = "MCP server for HTTP requests with browser-based authentication"
authors = [{name = "User", email = "user@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.25.0",
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
web-auth-mcp = "web_auth_mcp.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
